Teacher's code for how to do the glitch art requierments for the 1st section. And my own for 2nd and 3rd section.
Currently nothing except the given code works. The 2nd section is trying to create sections of the image which will have no green value and randomly assign them around. The 3rd section needs to add a bobRoss.png if a color is 100 away from orange as well as mirror 50-100 pixels.