Created a new method called classFIlter as well as added a BabRoss.png image for the class filter assignment.
The code in the classFilter is not currently correcct but the aim is to try to get the origional image to replace colors that are less than 180 from the color orange to the colors in the BobRoss.png.

Currently there is an out of bounds expetion for the array at line 281. Likely due to the BobRoss image array amount.